feat: Protocol updates: Support stateless transactions (#415)
# Protocol updates

This PR makes some clarifications around several concepts and intends to
make simple 'chit-chat' like scenarios, or stateless interactions,
easier. In particular:

- Introduces a messageId that is created by the message creator (both
client and server)
- Clarifies the ownership of a taskId. This is always determined by the
server (if/when a task is needed). Client should send follow up messages
with the relevant taskId to continue interactions around an existing
task
- Changes the name 'sessionId' to 'contextId' and allows it to be
attached to messages and tasks. The server is responsible for creating
the contextId and the client is responsible for using it to continue
interactions with shared context.
- Introduces a new message/send endpoint. This is parallel to the
task/send endpoint (which is deprecated as of this PR and will be
removed in a future PR). This endpoint returns either a Message object
(if there is not state/task to track) or a task object. If the request
is blocking, it will only return when the task is in a suspended state.
If the request is non-blocking, it will return the status of the initial
task creation. The client can then poll for task updates.
- Introduces a new message/stream endpoint. This is parallel to the
task/sendSubscribe endpoint (which is deprecated as of this PR and will
be removed in a future PR). The endpoint has the following expectations:
- If the response is a simple Message with no statefulness, returns a
Message and the channel should be closed.
- Otherwise, the first thing returned is an initial task object with the
'submitted' state. Each subsequent event is a TaskStatusUpdateEvent or
TaskArtifactUpdateEvent until the final TaskStatusUpdateEvent is
received.

The demo UI app and the cli app have been updated to use the new
endpoints, and no longer use the old ones. The biggest delta is that the
new endpoints create the task and context identifiers when the remote
agent deems it necessary, and the client is not responsible (nor should)
generate these itself. Also, all messages are required to have a
messageId with a value created by the sender. To continue a multi-turn
interaction the same contextId must be used in subsequent messages send
from the client to the server. To continue an interaction within the
scope of a task, the contextId and taskId must be included in the
message sent from the client.

---------

Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>